Avoid some dynamic memory allocation in parseOverlap()
We still generate some strings, but at least
not the format string itself.

Note that we use "cstring" to get around Nim's thread-safety
checker. A Nim string is dynamic, so it would not be technically
thread-safe. But we know it is constant for the duration of the
program, so the cstring pointer is perfectly safe.